% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_routes}
\alias{batch_routes}
\title{Interface to CycleStreets Batch Routing API}
\usage{
batch_routes(
  desire_lines,
  name = "test batch",
  serverId = 21,
  strategies = "quietest",
  bothDirections = 1,
  minDistance = 50,
  maxDistance = 5000,
  filename = "test",
  includeJsonOutput = 1,
  emailOnCompletion = "you@example.com",
  username = "yourname",
  password = Sys.getenv("CYCLESTREETS_PW"),
  base_url = "https://api.cyclestreets.net/v2/batchroutes.createjob",
  id = 1
)
}
\arguments{
\item{desire_lines}{Geographic desire lines representing origin-destination data}

\item{name}{The name of the batch routing job for CycleStreets}

\item{serverId}{The server ID to use (21 by default)}

\item{strategies}{Route plan types, e.g. \code{"fastest"}}

\item{bothDirections}{int (1|0)
Whether to plan in both directions, i.e. A-B as well as B-A.}

\item{minDistance}{Min distance}

\item{maxDistance}{Max distance}

\item{filename}{Character string}

\item{includeJsonOutput}{int (1|0)
Whether to include a column in the resulting CSV data giving the full JSON output from the API, rather than just summary
information like distance and time.}

\item{emailOnCompletion}{Email on completion?}

\item{username}{string
Your CycleStreets account username. In due course this will be replaced with an OAuth token.}

\item{password}{string
Your CycleStreets account password.}

\item{base_url}{The base url from which to construct API requests
(with default set to main server)}

\item{id}{int
Batch job ID, as returned from batchroutes.createjob.
action string (start|pause|continue|terminate)
Action to take. Available actions are:
start: Start (open) job
pause: Pause job
continue: Continue (re-open) job
terminate: Terminate job and delete data}
}
\description{
Interface to CycleStreets Batch Routing API
}
