% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclestreets_column_names}
\alias{cyclestreets_column_names}
\title{Prices of 50,000 round cut diamonds.}
\format{
An object of class \code{character} of length 44.
}
\source{
\url{www.cyclestreets.net}
}
\usage{
cyclestreets_column_names
}
\description{
Variables provided by CycleStreets in their journey data
}
\keyword{datasets}
